create test.c for testing implementation
